requirements.txt
discord.py>=2.0.0
aiomysql
python-dotenv
chat-exporter